[Ceby] : Add user key mapping functionality and related classes
This commit introduces a new record `UserKeyMappingData` and updates `AppSettings` to include `UserKeyMappingOptions` and `UserKeyMappingConfig`. New interfaces `ICreateUserDetailsToStorageCommand` and `IGetUserDetailsFromStorageQuery` are created for handling user key data operations. The `CreateUserDetailsToStorageCommand` and `GetUserDetailsFromStorageQuery` classes are updated to implement these interfaces, enabling the creation and retrieval of user key data in Azure Table Storage. Additionally, the project file is modified to include a new SDK declaration.